requirements.txt
tqdm
rotary_embedding_torch
transformers
tokenizers
inflect
progressbar
einops
unidecode
scipy
librosa
# numba==0.48.0
ffmpeg
